# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "這是課程學習網站的 API 文件。",
        "title": "課程學習網站 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/api/v1",
    "paths": {
        "/register": {
            "post": {
                "description": "註冊新用戶帳號",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用戶管理"
                ],
                "summary": "用戶註冊",
                "parameters": [
                    {
                        "description": "註冊請求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "註冊成功",
                        "schema": {
                            "$ref": "#/definitions/handler.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/{username}": {
            "get": {
                "description": "根據用戶名獲取用戶信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用戶管理"
                ],
                "summary": "獲取用戶信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用戶名",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "獲取成功",
                        "schema": {
                            "$ref": "#/definitions/handler.GetUserResponse"
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "用戶不存在",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "用戶名已存在"
                }
            }
        },
        "handler.GetUserResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "獲取用戶信息成功"
                },
                "user": {
                    "$ref": "#/definitions/handler.UserInfo"
                }
            }
        },
        "handler.RegisterRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "nickname": {
                    "type": "string",
                    "example": "測試用戶"
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                },
                "username": {
                    "type": "string",
                    "example": "testuser"
                }
            }
        },
        "handler.RegisterResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "註冊成功"
                },
                "user": {
                    "$ref": "#/definitions/handler.UserInfo"
                }
            }
        },
        "handler.UserInfo": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string",
                    "example": ""
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "nickname": {
                    "type": "string",
                    "example": "測試用戶"
                },
                "username": {
                    "type": "string",
                    "example": "testuser"
                }
            }
        }
    }
}